12/23 :truck: : user api와 register api 분리
